requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
asgiref==3.8.1
black==24.10.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
clickhouse-toolset==0.33.dev0
colorama==0.4.6
croniter==1.3.8
cryptography==41.0.7
Django==5.1.4
djangorestframework==3.15.2
gitdb==4.0.11
GitPython==3.1.43
humanfriendly==8.2
idna==3.10
inflection==0.5.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==2.6.2
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyperclip==1.8.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
shandy-sqlfmt==0.11.1
six==1.17.0
smmap==5.0.1
sqlparse==0.5.3
tinybird-cli==5.13.1
toposort==1.10
tornado==6.0.4
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.20
wheel==0.45.1
